Add AuthenticationConcern to group auth logic
Just to keep ApplicationController a bit tidier.
